% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pu_labels.R
\name{add_pu_labels}
\alias{add_pu_labels}
\title{Add PU Labels}
\usage{
add_pu_labels(df, pu.pi)
}
\arguments{
\item{df}{A dataframe containing data from the HiggsML challenge (must have true labels of "s" and "b" for signal and background stored in a column named "Label").}

\item{pu.pi}{The proportion of background samples to maintain their correct labels.}
}
\value{
The dataframe df with an extra column containing the PU labels. The column will be named "pu[pu.pi]" where [pu.pi] represents the numeric value of [pu.pi] as a string.
}
\description{
To simulate positive and unlabelled (PU) data we denote the background samples as positive with label 1 and signal samples as negative with label 0, before misclassifying some proportion (1-pu.pi) of the background samples.
This function adds the resulting labels as columns in the dataset it is passed as an argument.
}
\examples{
\dontrun{
  add_pu_labels(training_dataset, 0.6)
}

}
