% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ams.R
\name{calculate_ams}
\alias{calculate_ams}
\title{Calculate Approximate Median Significance (AMS)}
\usage{
calculate_ams(
  y,
  yhat,
  w,
  original_sum_b = 125317.9,
  original_sum_s = 212.418,
  b_r = 10
)
}
\arguments{
\item{y}{The true class labels (signal = 0, background = 1).}

\item{yhat}{The predicted class labels (signal = 0, background = 1).}

\item{w}{Weights of the data points in question (as provided in the original dataset).}

\item{original_sum_b}{Sum of background sample weights. Default is 125317.9 to match the Kaggle training set.}

\item{original_sum_s}{Sum of signal sample weights. Default is 212.418 to match the Kaggle training set.}

\item{b_r}{Regularisation term. Default is 10 as per the HiggsML Challenge document.}
}
\value{
The AMS of the predicted class labels given the true values.
}
\description{
This formula is from: \url{http://opendata.cern.ch/record/328}.
}
\examples{
y = rep(0, 5)
w = rep(1, 5)
calculate_ams(y, y, w)

}
