% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ams.R
\name{calculate_ams_unormalized}
\alias{calculate_ams_unormalized}
\title{Calculate Approximate Median Significance (AMS) Without Sample Weight Normalisation}
\usage{
calculate_ams_unormalized(y, yhat, w, b_r = 10)
}
\arguments{
\item{y}{The true class labels (signal = 0, background = 1).}

\item{yhat}{The predicted class labels (signal = 0, background = 1).}

\item{w}{Weights of the data points in question (as provided in the original dataset).}

\item{b_r}{Regularisation term. Default is 10 as per the HiggsML Challenge document.}
}
\value{
The AMS of the predicted class labels given the true values.
}
\description{
This formula is from: \url{http://opendata.cern.ch/record/328}.
}
\examples{
y = rep(0, 5)
w = rep(1, 5)
calculate_ams(y, y, w)

}
