% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_features.R
\name{create_new_features}
\alias{create_new_features}
\title{Create New Features}
\usage{
create_new_features(df)
}
\arguments{
\item{df}{A dataframe containing data from the HiggsML challenge. Must contain all primitive (PRI_) features as columns.}
}
\value{
The same dataframe given as an argument but with columns added for each of the extra features.
}
\description{
To improve classification performance, we can calculate these extra features using the provided primitive features.
Each of these new features has a physical meaning that will hopefully be useful to the classification algorithms.
}
\examples{
\dontrun{
  create_new_features(training_dataset)
}
}
